+++ lib/shaped/array.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-class Shaped::Array
+class Shaped::Array < Shaped::Shape
+  include Shaped::ExpectationChecking
+
   def initialize(shape_description)
     unless shape_description.size == 1
       raise(Shaped::InvalidShapeDescription, <<~ERROR.strip)
@@ -8,17 +10,23 @@ class Shaped::Array
       ERROR
     end
 
-    unless shape_description.all? { _1.is_a?(Class) }
-      raise(Shaped::InvalidShapeDescription, <<~ERROR.strip)
-        The element of a Shaped::Array description must be a class (ex: `[String]`).
-      ERROR
+    unless shape_description.all? { _1.is_a?(Class) || _1.is_a?(Shaped::Shape) }
+      raise(
+        Shaped::InvalidShapeDescription,
+        'The element of a Shaped::Array description must be a class (ex: `[String]`) ' \
+        'or a Shaped::Shape (ex.: `[Shaped::Hash(email: String)]`).',
+      )
     end
 
     @shape_klass = shape_description.first
     @match_failure_reasons = {}
   end
 
-  def matched_by?(array)
+  def descriptor
+    "Array shaped like [#{@shape_klass}]"
+  end
+
+  def matched_by?(array, prepended_path: [])
     # avoid a memory leak from indefinitely adding to the @match_failure_reasons hash; we just want
     # to use it to memoize recently checked arrays (we'll somewhat arbitrarily go with 10 of them)
     if @match_failure_reasons.size > 10
@@ -32,14 +40,14 @@ class Shaped::Array
     else
       all_elements_match = true
       array.each_with_index do |element, index|
-        next if element.is_a?(@shape_klass)
+        next if meets_expectation?(element, @shape_klass, prepended_path: prepended_path + [index])
 
         all_elements_match = false
-        @match_failure_reasons[array] =
+        @match_failure_reasons[array] ||=
           Shaped::MatchFailureReason.new(
-            path: [index],
-            expected: @shape_klass,
-            actual: element.class,
+            path: prepended_path + [index],
+            expected_klass: @shape_klass,
+            actual_value: element,
           )
         break
       end