+++ lib/shaped/expected_class_descriptions.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Shaped::ExpectedClassDescriptions
+  private
+
+  def expected_class_descriptor(expected_klass)
+    if expected_klass.is_a?(Shaped::Shape)
+      expected_klass.descriptor
+    else
+      expected_klass.name
+    end
+  end
+end