+++ lib/shaped/shape.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+# This empty class just exists so that both `Shaped::Array` and `Shaped::Hash` can have it as a
+# common ancestor.
+class Shaped::Shape
+end